+++ martini/legacy/ajax/customer_soa_results_function.php
@@ -5,21 +5,21 @@ function get_customer_soa_results($customer_id,$adv)
     $ret = [];
     while($picksheet = mysqli_fetch_assoc($customerPicksheets)){
 
-        $picksheet['credit'] = (float) round(totalValueCreditedOnInvoiceID($picksheet['id']),2,PHP_ROUND_HALF_DOWN);
-        $picksheet['price'] = (float) round(invoiceTotal($picksheet['id']),2,PHP_ROUND_HALF_DOWN);
+        $picksheet['credit'] = (double) round(totalValueCreditedOnInvoiceID($picksheet['id']),2,PHP_ROUND_HALF_DOWN);
+        $picksheet['price'] = (double) round(invoiceTotal($picksheet['id']),2,PHP_ROUND_HALF_DOWN);
 
         $picksheet['date'] = str_replace('/', '-', $picksheet['estimated_delivery_date']);
         $picksheet['datetime'] = strtotime($picksheet['date']);
         $picksheet['date'] = date('d/m/Y', $picksheet['datetime']);
 
-	    $picksheet['paid'] = (float) round($picksheet['paid'],2,PHP_ROUND_HALF_DOWN);
+	    $picksheet['paid'] = (double) round($picksheet['paid'],2,PHP_ROUND_HALF_DOWN);
         $picksheet['invoicePaid'] = false;
         $epsilon = 0.00001;
 
         if(($picksheet['price'] - $picksheet['paid']) <= $epsilon){
             $picksheet['invoicePaid'] = true;
         }
-	    $picksheet['outstanding'] = round((float) $picksheet['price'] - $picksheet['paid'] - $picksheet['credit'],2,PHP_ROUND_HALF_DOWN);
+	    $picksheet['outstanding'] = round((double) $picksheet['price'] - $picksheet['paid'] - $picksheet['credit'],2,PHP_ROUND_HALF_DOWN);
         if ($adv == true && ($picksheet['outstanding'] > -0.02 && $picksheet['outstanding'] < 0.02)) continue;
         $picksheet['credited'] = totalValueCreditedOnInvoiceID($picksheet['id']);
         $picksheet['hasReturns'] = doesInvoiceHaveReturns($picksheet['id']);
@@ -99,7 +99,7 @@ function check_customer_outstanding_cache($customer_id,$forceReload = false)
     $lastRow = null;
     while($row = mysqli_fetch_assoc($checkQ))
     {  
-        $row['outstanding'] = (float)(getOutstandingPicksheetTotal($row['id']) - totalValueCreditedOnInvoiceID($row['id']));
+        $row['outstanding'] = (double)(getOutstandingPicksheetTotal($row['id']) - totalValueCreditedOnInvoiceID($row['id']));
         if ($row['outstanding'] > 0)
         {
             $cacheRow['pending'] = $row;           
@@ -150,11 +150,11 @@ function update_customer_outstanding_cache($cacheRow)
     
     if ($cacheRow['newRow'] == true)
     {
-        $sql = "INSERT INTO customer_outstanding_cache (`customer_id`, `pickersheet_id`, `invoice_payment_id`, `oldest_unpaid_id`, `outstanding`) VALUES (".$cacheRow['customer_id'].",".$cacheRow['pickersheet_id'].",".$cacheRow['invoice_payment_id'].",'".$cacheRow['oldest_unpaid_id']."','".(float)$cacheRow['outstanding']."')";
+        $sql = "INSERT INTO customer_outstanding_cache (`customer_id`, `pickersheet_id`, `invoice_payment_id`, `oldest_unpaid_id`, `outstanding`) VALUES (".$cacheRow['customer_id'].",".$cacheRow['pickersheet_id'].",".$cacheRow['invoice_payment_id'].",'".$cacheRow['oldest_unpaid_id']."','".(double)$cacheRow['outstanding']."')";
     }
     else
     {
-        $sql = "UPDATE customer_outstanding_cache SET `pickersheet_id` = ".$cacheRow['pickersheet_id'].", `invoice_payment_id` = ".$cacheRow['invoice_payment_id'].", `outstanding` = '".(float)$cacheRow['outstanding']."', `oldest_unpaid_id` = '".$cacheRow['oldest_unpaid_id']."' WHERE `customer_outstanding_cache`.`customer_id` = ".$cacheRow['customer_id'];
+        $sql = "UPDATE customer_outstanding_cache SET `pickersheet_id` = ".$cacheRow['pickersheet_id'].", `invoice_payment_id` = ".$cacheRow['invoice_payment_id'].", `outstanding` = '".(double)$cacheRow['outstanding']."', `oldest_unpaid_id` = '".$cacheRow['oldest_unpaid_id']."' WHERE `customer_outstanding_cache`.`customer_id` = ".$cacheRow['customer_id'];
     }
     $x = prepareExecuteQuery($sql);
 }
