+++ martini/legacy/exportStockPDF.php
@@ -311,15 +311,15 @@ INNER JOIN `weights` ON product.id = weights.product_id
                 $temp_weight = totalWeightOfProduct($product2_productids);
             }
 
-            $html .= '<td class="cell unit">' . number_format((float)$temp_weight, 3, '.', '') . '</td>';
+            $html .= '<td class="cell unit">' . number_format((double)$temp_weight, 3, '.', '') . '</td>';
             $total_weight += $temp_weight;
         }
 
-        $total_cost += number_format((float)$productsRow['cost'], 2, '.', '');
-        $total_price += number_format((float)$productsRow['price'], 2, '.', '');
+        $total_cost += number_format((double)$productsRow['cost'], 2, '.', '');
+        $total_price += number_format((double)$productsRow['price'], 2, '.', '');
 
-        $html .= '<td class="cell cost">£' . number_format((float)$productsRow['cost'], 2, '.', '') . '</td>';
-        $html .= '<td class="cell price">£' .  number_format((float)$productsRow['price'], 2, '.', '') . '</td>';
+        $html .= '<td class="cell cost">£' . number_format((double)$productsRow['cost'], 2, '.', '') . '</td>';
+        $html .= '<td class="cell price">£' .  number_format((double)$productsRow['price'], 2, '.', '') . '</td>';
         $html .='</tr>';
     }
     $html .= '<tr>';
@@ -333,7 +333,7 @@ INNER JOIN `weights` ON product.id = weights.product_id
     $html .= '<td class="cell brand"></td>';
     $html .= '<td class="cell daterange"></td>';
     $html .= '<td class="cell ppc"></td>';
-    $html .= '<td class="cell unit">'. number_format((float)$total_weight, 3, '.', ',') .'</td>';
+    $html .= '<td class="cell unit">'. number_format((double)$total_weight, 3, '.', ',') .'</td>';
     $html .= '<td class="cell cost"><b style="font-size:8px;">£' . $total_cost . '</b></td>';
     $html .= '<td class="cell price"><b style="font-size:8px;">£' . $total_price . '</b></td>';
     $html .='</tr></table>';
